+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-* 1.0.0-next (@release_date@)
+* 1.0.0-beta3 (2015-03-19)
 
  * replaced `Assert` by webmozart/assert
  * renamed `ResourceDiscovery::find()` to `findByType()`